+++ vortex-duckdb/cbindgen.toml
@@ -1,4 +1,5 @@
 language = "C"
+cpp_compat = true
 
 header = """
 // SPDX-License-Identifier: Apache-2.0
@@ -23,3 +24,6 @@ trailer = """
 
 // clang-format on
 """
+
+[fn]
+prefix = "extern"
