+++ vortex-duckdb/include/vortex.h
@@ -14,6 +14,10 @@ extern "C" {
 #include "duckdb.h"
 
 
+#ifdef __cplusplus
+extern "C" {
+#endif // __cplusplus
+
 /**
  * Global symbol visibility in the Vortex extension:
  * - Rust functions use C ABI with "_rust" suffix (e.g., vortex_init_rust)
@@ -25,20 +29,69 @@ extern "C" {
  *
  * The DuckDB extension ABI initialization function.
  */
-void vortex_init_rust(duckdb_database db);
+extern void vortex_init_rust(duckdb_database db);
 
 /**
  * The DuckDB extension ABI version function.
  * This function returns the version of the DuckDB library the extension is built against.
  */
-const char *vortex_version_rust(void);
+extern const char *vortex_version_rust(void);
 
 /**
  * An additional function we export to expose the version of the extension itself to C++ code.
  */
-const char *vortex_extension_version_rust(void);
+extern const char *vortex_extension_version_rust(void);
 
-bool duckdb_vx_pushdown_expression(duckdb_vx_expr expr);
+extern void duckdb_table_function_to_string(void *bind_data, duckdb_vx_string_map map);
+
+extern
+bool duckdb_table_function_statistics(const void *bind_data,
+                                      uintptr_t column_index,
+                                      duckdb_column_statistics *stats_out);
+
+extern double duckdb_table_function_scan_progress(void *global_state);
+
+extern
+void duckdb_table_function_get_partition_data(void *global_init_data,
+                                              void *local_init_data,
+                                              duckdb_vx_partition_data *partition_data_out);
+
+extern
+bool duckdb_table_function_pushdown_complex_filter(void *bind_data,
+                                                   duckdb_vx_expr expr,
+                                                   duckdb_vx_error *error_out);
+
+extern
+void duckdb_table_function_scan(void *global_init_data,
+                                void *local_init_data,
+                                duckdb_data_chunk output,
+                                duckdb_vx_error *error_out);
+
+extern bool duckdb_table_function_pushdown_expression(duckdb_vx_expr expr);
+
+extern
+void duckdb_table_function_cardinality(void *bind_data,
+                                       duckdb_vx_node_statistics *node_stats_out);
+
+extern
+duckdb_vx_data duckdb_table_function_init_global(const duckdb_vx_tfunc_init_input *init_input,
+                                                 duckdb_vx_error *error_out);
+
+extern duckdb_vx_data duckdb_table_function_init_local(void *global_init_data);
+
+extern
+duckdb_vx_data duckdb_table_function_bind(duckdb_client_context ctx,
+                                          duckdb_vx_tfunc_bind_input bind_input,
+                                          duckdb_vx_tfunc_bind_result bind_result,
+                                          duckdb_vx_error *error_out);
+
+extern
+duckdb_vx_data duckdb_table_function_bind_data_clone(const void *bind_data,
+                                                     duckdb_vx_error *error_out);
+
+#ifdef __cplusplus
+}  // extern "C"
+#endif  // __cplusplus
 
 #ifdef __cplusplus
 }
