+++ vortex-duckdb/src/duckdb/database.rs
@@ -131,6 +131,14 @@ impl Database {
 }
 
 impl DatabaseRef {
+    pub fn register_table_functions(&self) -> VortexResult<()> {
+        duckdb_try!(
+            unsafe { cpp::duckdb_vx_register_table_functions(self.as_ptr()) },
+            "Failed to register table functions"
+        );
+        Ok(())
+    }
+
     /// Connects to the DuckDB database.
     pub fn connect(&self) -> VortexResult<Connection> {
         Connection::connect(self)
